# docs/swagger.yaml
definitions:
...
  dto.OrderCreateRequestDto:
    properties:
      key:
        type: string
      pickup_schedule:
        type: string
    required:
    - key
    - pickup_schedule
    type: object
  dto.OrderCreateResponseDto:
    properties:
      order_id:
        type: string
    required:
    - order_id
    type: object
  dto.OrderFindResponseDto:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.PaginationMetaDto'
    type: object
  dto.OrderResponseDto:
    properties:
      created_at:
        type: string
      item:
        $ref: '#/definitions/models.OrderItem'
      librarian_id:
        type: string
      order_id:
        type: string
      pickup_schedule:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.UserFindResponseDto:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.PaginationMetaDto'
    type: object
...
  dto.UserRegisterRequestDto:
    properties:
      email:
        maxLength: 60
        type: string
      first_name:
        maxLength: 30
        type: string
      last_name:
        maxLength: 30
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
  dto.UserRegisterResponseDto:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  models.OrderItem:
    properties:
      cover_id:
        type: number
      edition_count:
        type: string
      key:
        type: string
      name:
        type: string
    type: object
  utils.PaginationMetaDto:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
    type: object
info:
  contact:
    email: djourdan555@gmail.com
    name: Dustin Jourdan
    url: https://github.com/dinorain
paths:
...
  /order:
    get:
      consumes:
      - application/json
      description: Find all orders
      parameters:
      - description: pagination size
        in: query
        name: size
        type: string
      - description: pagination page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderFindResponseDto'
      security:
      - ApiKeyAuth: []
      summary: Find all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Admin create order
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.OrderCreateRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderCreateResponseDto'
      security:
      - ApiKeyAuth: []
      summary: To register order
      tags:
      - Orders
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Find existing order by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponseDto'
      security:
      - ApiKeyAuth: []
      summary: Find order
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Librarian accept order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponseDto'
      security:
      - ApiKeyAuth: []
      summary: Accept order
      tags:
      - Orders
  /user:
    get:
      consumes:
      - application/json
      description: Admin find all users
      parameters:
      - description: pagination size
        in: query
        name: size
        type: string
      - description: pagination page
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserFindResponseDto'
      security:
      - ApiKeyAuth: []
      summary: Find all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Admin create user
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserRegisterResponseDto'
      security:
      - ApiKeyAuth: []
      summary: To register user
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"